feat(publicURL): #28 Moved the saveUserDetails functionality into user service
